refactor: split library API out of __main__, replace user-facing asserts
Two structural cleanups Fable's outside review flagged:

1. **API in `__main__.py`**: importing `invoice2data` pulled in click and
   the CLI's ANSI/log-formatter plumbing (927-line `__main__.py` mixing
   library and CLI concerns). Move `extract_data`, `Invoice2Data`, the
   backend-cascade helpers, and `DEFAULT_INPUT_READERS` to a new
   `invoice2data.api` module. `__main__` is now a thin click shell
   that imports from `api`. `__init__.py` re-exports from `api`.
   `__main__.py` also re-exports the moved helpers so pre-refactor
   callers (`from invoice2data.__main__ import extract_data`) keep
   working. Verified via `sys.modules`: `import invoice2data` no
   longer pulls in `click`.

2. **Assert-based validation** in a Production/Stable release: asserts
   vanish under `python -O` and expose `AssertionError` -- an internal
   invariant class -- to library callers. Add a
   `TemplateSyntaxError(InvoiceProcessingError, ValueError)` and
   convert the user-facing sites:
   - `invoice_template.parse_number`: decimal-separator misuse (repeated
     separator, non-string option) + non-string input
   - `invoice_template.coerce_type`: unknown field `type`
   - `invoice_template.prepare_input`: malformed `replace` pair
   - `extract.parsers.lines`: missing `line`/`start`/`end` settings
   - `extract.plugins.tables`: missing `start`/`end`/`body` settings
   Subclasses `ValueError` so the cascade's existing
   `except ValueError` retry keeps skipping broken templates gracefully.
   Area-detail asserts in the input backends (pdfium/pdftotext/tesseract)
   left as follow-up (they're user-facing but cross more file boundaries).

Test suite green modulo 9 pre-existing failures (missing `pypdfium2` in
local env + a couple of golden-fixture regexes and an unrelated gvision
attribute) all confirmed identical on unstashed `next`. One test
(`test_preferred_module_unavailable_is_ignored`) updated to patch
`invoice2data.api.is_available` -- that's where the symbol is now
looked up.

Draft: opens for review of the module split before landing.
